+++ lib/mailgun_rails.rb
@@ -2,7 +2,7 @@ require 'action_mailer'
 require 'json'
 
 
-Dir[File.dirname(__FILE__) + '/mailgun/*.rb'].each {|file| require file }
+Dir[File.dirname(__FILE__) + '/mailgun_rails/*.rb'].each {|file| require file }
 
-module Mailgun
+module MailgunRails
 end
