+++ lib/mailgun_rails/attachment.rb
@@ -1,4 +1,4 @@
-module Mailgun
+module MailgunRails
   class Attachment < StringIO
     attr_reader :original_filename, :content_type, :path
