+++ lib/mailgun_rails/mail_ext.rb
