+++ lib/mailgun_rails/version.rb
@@ -1,3 +1,3 @@
-module Mailgun
+module MailgunRails
   VERSION = "0.8.0"
 end
