+++ mailgun_rails.gemspec
@@ -1,12 +1,12 @@
 $:.push File.expand_path("../lib", __FILE__)
 
 # Maintain your gem's version:
-require "mailgun/version"
+require "mailgun_rails/version"
 
 # Describe your gem and declare its dependencies:
 Gem::Specification.new do |s|
   s.name        = "mailgun_rails"
-  s.version     = Mailgun::VERSION
+  s.version     = MailgunRails::VERSION
   s.authors     = ["Jorge Manrubia"]
   s.email       = ["jorge.manrubia@gmail.com"]
   s.homepage    = "https://github.com/jorgemanrubia/mailgun_rails/"
