+++ spec/lib/mailgun_rails/attachment_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
-require 'mailgun/attachment'
+require 'mailgun_rails/attachment'
 
-describe Mailgun::Attachment do
+describe MailgunRails::Attachment do
   describe "#send_message" do
     before do
       @mail = Mail.new()
@@ -12,7 +12,7 @@ describe Mailgun::Attachment do
     end
 
     it 'should respond to rest_client api' do
-      attachment = Mailgun::Attachment.new(@mail.attachments.first)
+      attachment = MailgunRails::Attachment.new(@mail.attachments.first)
       attachment.respond_to?(:path).should eq(true)
       attachment.respond_to?(:original_filename).should eq(true)
       attachment.respond_to?(:content_type).should eq(true)
@@ -20,17 +20,17 @@ describe Mailgun::Attachment do
     end
 
     it 'should set cid as original_filename' do
-      attachment = Mailgun::Attachment.new(@mail.attachments.inline.first, inline: true)
+      attachment = MailgunRails::Attachment.new(@mail.attachments.inline.first, inline: true)
       attachment.original_filename.should eq(@mail.attachments.inline.first.cid)
     end
 
     it 'should set filename as original_filename' do
-      attachment = Mailgun::Attachment.new(@mail.attachments.first)
+      attachment = MailgunRails::Attachment.new(@mail.attachments.first)
       attachment.original_filename.should eq(@mail.attachments.first.filename)
     end
 
     it 'should set filename as original_filename for hash' do
-      attachment = Mailgun::Attachment.new(@mail.attachments.last)
+      attachment = MailgunRails::Attachment.new(@mail.attachments.last)
       attachment.original_filename.should eq(@mail.attachments.last.filename)
     end
   end
